move icon picker
