+++ src/packages/core/property-editor/uis/manifests.ts
@@ -4,7 +4,7 @@ import { manifest as colorPicker } from '../../../property-editors/color-picker/
 import { manifest as datePicker } from '../../../property-editors/date-picker/manifests.js';
 import { manifest as dropdown } from '../../../property-editors/dropdown/manifests.js';
 import { manifest as eyeDropper } from '../../../property-editors/eye-dropper/manifests.js';
-import { manifest as iconPicker } from './icon-picker/manifests.js';
+import { manifest as iconPicker } from '../../../property-editors/icon-picker/manifests.js';
 import { manifest as label } from './label/manifests.js';
 import { manifest as multipleTextString } from './multiple-text-string/manifests.js';
 import { manifest as multiUrlPicker } from './multi-url-picker/manifests.js';
