+++ src/packages/property-editors/icon-picker/manifests.ts
