+++ src/packages/property-editors/icon-picker/property-editor-ui-icon-picker.element.ts
