+++ src/packages/property-editors/icon-picker/property-editor-ui-icon-picker.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, Story } from '@storybook/web-components';
-import type { UmbIconPickerModalElement } from '../../../modal/common/icon-picker/icon-picker-modal.element.js';
+import type { UmbIconPickerModalElement } from '../../core/modal/common/icon-picker/icon-picker-modal.element.js';
 import { html } from '@umbraco-cms/backoffice/external/lit';
 
 import './property-editor-ui-icon-picker.element.js';
