+++ src/packages/property-editors/icon-picker/property-editor-ui-icon-picker.test.ts
